edit and delete question with authentication

+++ SOAP-WS/src/org/tusiri/ws/question/Question.java
@@ -10,7 +10,6 @@ import java.sql.*;
 import org.apache.http.client.ClientProtocolException;
 import org.json.simple.parser.ParseException;
 import org.tusiri.ws.db.DBConnection;
-
 import org.tusiri.ws.token_validity.CheckTokenValidity;
 import org.tusiri.ws.token_validity.TokenValidity;
 
@@ -69,7 +68,7 @@ public class Question {
 	}
 	
 	@WebMethod
-	public int editQuestion(String access_token,String title,String content, int id_question) throws ClientProtocolException, IOException, ParseException{
+	public int editQuestion(String access_token,String title,String content, int id_question) throws ClientProtocolException, IOException, ParseException, SQLException{
 		int q_id = 0;
 		
 		System.out.println(access_token);
@@ -82,29 +81,39 @@ public class Question {
 			TokenValidity validity = checker.check();
 			
 			if(validity.getIsValid()){
-				//Masukkan ke database
 				DBConnection dbc = new DBConnection();
 				PreparedStatement stmt = dbc.getDBStmt();
 				Connection conn = dbc.getConn();
-				try{
-					String sql = "UPDATE question SET content = ?, topic = ?, question_date() = NOW() WHERE id_question = ?";
-					stmt = conn.prepareStatement(sql,Statement.RETURN_GENERATED_KEYS);
-					stmt.setString(1,content);
-					stmt.setString(2,title);
-					stmt.setInt(3, id_question);
-					stmt.executeUpdate();
-					ResultSet rs = stmt.getGeneratedKeys();
-		            while (rs.next()) {
-		               q_id = rs.getInt(1);
-		            } 	
-					System.out.println("q_id = " + q_id);
-					//res = 1;
-				} catch(SQLException se){
-					//Handle errors for JDBC
-					se.printStackTrace();
-				} catch(Exception e){
-					//Handle errors for Class.forName
-					e.printStackTrace();
+				
+				//Check apakah user_id sesuai dengan user_id question yang bersangkutan. 
+				String sqlCheck = "SELECT id_user FROM question WHERE id_question = ?";
+				stmt = conn.prepareStatement(sqlCheck);
+				stmt.setInt(1,id_question);
+				ResultSet rsCheck = stmt.executeQuery();
+				if (rsCheck.next()){
+					if(rsCheck.getInt("id_user") == validity.getIdUser()){//jika id sesuai
+						try{
+							//Update database
+							String sql = "UPDATE question SET content = ?, topic = ?, question_date = NOW() WHERE id_question = ?";
+							stmt = conn.prepareStatement(sql,Statement.RETURN_GENERATED_KEYS);
+							stmt.setString(1,content);
+							stmt.setString(2,title);
+							stmt.setInt(3, id_question);
+							stmt.executeUpdate();
+							ResultSet rs = stmt.getGeneratedKeys();
+				            while (rs.next()) {
+				               q_id = rs.getInt(1);
+				            } 	
+							System.out.println("q_id = " + q_id);
+							//res = 1;
+						} catch(SQLException se){
+							//Handle errors for JDBC
+							se.printStackTrace();
+						} catch(Exception e){
+							//Handle errors for Class.forName
+							e.printStackTrace();
+						}
+					}
 				}
 			}
 		} catch (MalformedURLException e) {
@@ -116,7 +125,7 @@ public class Question {
 	}
 	
 	@WebMethod
-	public int deleteQuestion(String access_token,int id_question) throws ClientProtocolException, IOException, ParseException{
+	public int deleteQuestion(String access_token,int id_question) throws ClientProtocolException, IOException, ParseException, SQLException{
 		int q_id = 0;
 		
 		System.out.println(access_token);
@@ -127,29 +136,40 @@ public class Question {
 			TokenValidity validity = checker.check();
 			
 			if(validity.getIsValid()){
-				//Masukkan ke database
 				DBConnection dbc = new DBConnection();
 				PreparedStatement stmt = dbc.getDBStmt();
 				Connection conn = dbc.getConn();
-				try{
-					String sql = "DELETE FROM question WHERE id_question = ?; "
-							+ "DELETE FROM answer WHERE id_question = ?";
-					stmt = conn.prepareStatement(sql,Statement.RETURN_GENERATED_KEYS);
-					stmt.setInt(1, id_question);
-					stmt.setInt(2, id_question);
-					stmt.executeUpdate();
-					ResultSet rs = stmt.getGeneratedKeys();
-		            while (rs.next()) {
-		               q_id = rs.getInt(1);
-		            } 	
-					System.out.println("q_id = " + q_id);
-					//res = 1;
-				} catch(SQLException se){
-					//Handle errors for JDBC
-					se.printStackTrace();
-				} catch(Exception e){
-					//Handle errors for Class.forName
-					e.printStackTrace();
+				
+				//Check apakah user_id sesuai dengan user_id question yang bersangkutan. 
+				String sqlCheck = "SELECT id_user FROM question WHERE id_question = ?";
+				stmt = conn.prepareStatement(sqlCheck);
+				stmt.setInt(1,id_question);
+				ResultSet rsCheck = stmt.executeQuery();
+				
+				if (rsCheck.next()){
+					if(rsCheck.getInt("id_user") == validity.getIdUser()){//jika id sesuai
+						//Delete question dan answer dari database
+						try{
+							String sql = "DELETE FROM question WHERE id_question = ?; "
+									+ "DELETE FROM answer WHERE id_question = ?";
+							stmt = conn.prepareStatement(sql,Statement.RETURN_GENERATED_KEYS);
+							stmt.setInt(1, id_question);
+							stmt.setInt(2, id_question);
+							stmt.executeUpdate();
+							ResultSet rs = stmt.getGeneratedKeys();
+				            while (rs.next()) {
+				               q_id = rs.getInt(1);
+				            } 	
+							System.out.println("q_id = " + q_id);
+							//res = 1;
+						} catch(SQLException se){
+							//Handle errors for JDBC
+							se.printStackTrace();
+						} catch(Exception e){
+							//Handle errors for Class.forName
+							e.printStackTrace();
+						}
+					}
 				}
 			}
 		} catch (MalformedURLException e) {
